implemented encoder direction flip
it was found found that V3.0 board had motor pins on M1 and M2 flipped between the EV3 port and the motor pins broken out. as a result, run target does not work correctly dueto the motor direction inconsistent with the increment/decrement of the encoder. flipEncoderDirection() is created to temporarily solve the problem. it is also a feature that would be useful in the event that the wires of the motor and encoders are not able to be flipped